tsmicroctl: allow for issuing sleep command later in system shutdown
As this binary is intended to run in many different environments
and platforms, it relies on information exposed in /sys to gather
compatibility information that is compared to board_t definitions
for each compatible platform.

However, for sleep, we may be running in an environment where /sys
and other kernel exported virtual filesystems are no longer available.
In this state, we use a generic platform definition. All platforms that
support this binary have their microcontroller on I2C bus 0, and at
address 0x54. Meaning we can still run in all of the valid environments
and still issue a sleep command.

No other commands are accepted in this state of operation. And
even then, running this command on unsupported platforms in the
state where /sys is unmounted or on a platform that does not
use device-trees can cause serious issues!

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>